requirements.txt
ultralytics
flask
requests
pillow
torch
torchvision
